Simplification of code structure
