+++ src/Primitives.cpp
@@ -1,5 +1,5 @@
-#include <DCA/Utils/ExpCoords.h>
-#include <DCA/Utils/Primitives.h>
+#include <DCA/ExpCoords.h>
+#include <DCA/Primitives.h>
 
 namespace DCA {
 
@@ -9,7 +9,7 @@ Vector3d Primitive::get_center_point() const {
     return compute_P(s, t);
 }
 
-Primitive::Primitive(const std::string& description, const double& safetyMargin) : Opt::FiniteDifference(description), safetyMargin(safetyMargin) {
+Primitive::Primitive(const std::string& description, const double& safetyMargin) : FiniteDifference(description), safetyMargin(safetyMargin) {
     if (safetyMargin < 0.0)
         throw std::runtime_error("Error in Primitive::Primitive -> invalid safety margin");
 }
@@ -74,93 +74,6 @@ void Primitive::check_t(const VectorXd& t) const {
 
 VectorXd Primitive::get_t_forFD() const {
     return VectorXd::Random(SIZE_T());
-}
-
-void Primitive::estimateTensor(TensorXd& tensor, const Function_Matrix& evaluate, const VectorXd& s, const VectorXd& t, D_WRT d_wrt, int firstDim,
-                               int secondDim) const {
-    VectorXd p_der = (d_wrt == D_S) ? s : t;
-    tensor.clear();
-
-    auto eval = [&](MatrixXd& mat, D_WRT d_wrt) -> void {
-        if (d_wrt == D_S)
-            evaluate(mat, p_der, t);
-        else if (d_wrt == D_T)
-            evaluate(mat, s, p_der);
-    };
-
-    MatrixXd f_P(firstDim, secondDim), f_M(firstDim, secondDim);
-    for (int i = 0; i < p_der.size(); i++) {
-        f_P.setZero();
-        f_M.setZero();
-
-        p_der[i] += deltaFD;
-        eval(f_P, d_wrt);
-
-        p_der[i] -= 2.0 * deltaFD;
-        eval(f_M, d_wrt);
-
-        p_der[i] += deltaFD;
-        tensor.emplace_back((f_P - f_M) / (2 * deltaFD));
-    }
-}
-
-void Primitive::testTensor(const Function_Matrix& evaluate, const Function_Tensor& analytic, const VectorXd& s, const VectorXd& t, D_WRT d_wrt,
-                           const std::string& name, int firstDim, int secondDim) const {
-    VectorXd p_der = (d_wrt == D_S) ? s : t;
-    uint pSize = (uint)p_der.size();
-
-    TensorXd fdTensor(pSize, MatrixXd::Zero(firstDim, secondDim));
-    TensorXd anaTensor(pSize, MatrixXd::Zero(firstDim, secondDim));
-
-    setFDCheckIsBeingApplied(true);
-    estimateTensor(fdTensor, evaluate, s, t, d_wrt, firstDim, secondDim);
-    analytic(anaTensor, s, t);
-    setFDCheckIsBeingApplied(false);
-
-    printFDCheck(fdTensor, anaTensor, name);
-}
-
-void Primitive::printFDCheck(const TensorXd& estimate, const TensorXd& analytic, const std::string& name) const {
-    bool checkSuccessful = true;
-    Logger::print(Logger::BLUE,
-                  "------------------------------------------------------------"
-                  "-----------------------------------------------\n");
-    Logger::print(Logger::DEFAULT, "Test ");
-    Logger::print(Logger::CYAN, "%s", name.c_str());
-    Logger::print(Logger::DEFAULT, " of ");
-    Logger::print(Logger::YELLOW, "%s", description.c_str());
-    if (analytic.size() > 0 && estimate.size() > 0) {
-        double norm_ana = 0.0, norm_est = 0.0;
-        for (int k = 0; k < (int)estimate.size(); k++) {
-            norm_ana += analytic[k].norm();
-            norm_est += estimate[k].norm();
-        }
-        Logger::print(Logger::DEFAULT, " with FD -> norms: Ana: %lf, FD: %lf\n", norm_ana, norm_est);
-    } else {
-        Logger::print(Logger::DEFAULT, " with FD -> norms: Ana: %lf, FD: %lf\n", 0.0, 0.0);
-    }
-
-    for (int k = 0; k < (int)estimate.size(); k++)
-        for (int i = 0; i < (int)estimate[k].rows(); i++) {
-            for (int j = 0; j < (int)estimate[k].cols(); j++) {
-                double absErr = std::abs(estimate[k].coeff(i, j) - analytic[k].coeff(i, j));
-                double relError = 2 * absErr / (eps + std::abs(estimate[k].coeff(i, j)) + std::abs(analytic[k].coeff(i, j)));
-                if (relError > relTol && absErr > absTol) {
-                    checkSuccessful = false;
-                    Logger::print(Logger::RED, "   MISMATCH");
-                    Logger::print(Logger::DEFAULT,
-                                  " -> Element: (%d, %d, %d), Ana val: %lf, FD val: "
-                                  "%lf. Error: %lf(%lf%%)\n",
-                                  i, j, k, analytic[k].coeff(i, j), estimate[k].coeff(i, j), absErr, relError * 100);
-                }
-            }
-        }
-    if (checkSuccessful)
-        Logger::print(Logger::GREEN, "   PASSED\n");
-    Logger::print(Logger::DEFAULT, "End of check of %s\n", description.c_str());
-    Logger::print(Logger::BLUE,
-                  "------------------------------------------------------------"
-                  "-----------------------------------------------\n");
 }
 
 //-----------------------------------------------------------------------------------------------------------------------------------------------------
