+++ src/Solver.cpp
@@ -1,18 +1,17 @@
-#include <DCA/Interactions/Solver.h>
-#include <DCA/Utils/Logger.h>
+#include <DCA/Logger.h>
+#include <DCA/Solver.h>
 
 namespace DCA {
-namespace Interactions {
 
 Solver::Solver(primitive_t primitive_A, primitive_t primitive_B)
-    : Opt::FiniteDifference(DCAObjective::get_primitives_description(primitive_A, primitive_B) + "-Solver"), objective(primitive_A, primitive_B) {
-    deltaFD = 1e-8;
-    absTol = 1e-4;
-}
+    : FiniteDifference(Objective::get_primitives_description(primitive_A, primitive_B) + "-Solver"), objective(primitive_A, primitive_B) {}
 
 void Solver::compute_t(VectorXd& t, const VectorXd& s, bool forFD) const {
     if ((int)t.size() != objective.get_total_size_t())
         throw std::runtime_error("Solver::compute_t: input size of t is not correct!");
+
+    if (t.size() == 0)
+        return;
 
     optimizer.solverResidual = forFD ? 1e-12 : 1e-6;
     optimizer.optimize(t, s, objective, 1000);
@@ -20,8 +19,8 @@ void Solver::compute_t(VectorXd& t, const VectorXd& s, bool forFD) const {
 
 void Solver::compute_dtds(MatrixXd& dtds, const VectorXd& s, const VectorXd& t) const {
     MatrixXd p2OpT2(t.size(), t.size()), p2OpTpS(t.size(), s.size());
-    objective.get_p2OpT2(p2OpT2, s, t);
-    objective.get_p2OpTpS(p2OpTpS, s, t);
+    objective.compute_p2OpT2(p2OpT2, s, t);
+    objective.compute_p2OpTpS(p2OpTpS, s, t);
 
     //Solve linear system
     Eigen::LLT<Eigen::MatrixXd, Eigen::Lower> solver;
@@ -99,7 +98,7 @@ bool Solver::are_t_and_s_in_sync(const VectorXd& s, const VectorXd& t) const {
         throw std::runtime_error("Solver::are_t_and_s_in_sync: input size of t is not correct!");
 
     VectorXd grad(t.size());
-    objective.get_pOpT(grad, s, t);
+    objective.compute_pOpT(grad, s, t);
     if (grad.norm() > 1e-6) {
         Logger::print(Logger::RED, "ERROR in Solver::are_t_and_s_in_sync -> s and t are NOT in sync! Gradient norm is: %lf", grad.norm());
         throw std::runtime_error("Solver::are_t_and_s_in_sync -> s and t are NOT in sync!");
@@ -142,5 +141,4 @@ void Solver::test_derivatives() const {
     //test_d2DdS2_WithFD(s, t);
 }
 
-}  // namespace Interactions
 }  // namespace DCA
